+++ EmmerceAPIHCMUE/Controllers/VoucherController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
+using System.Data;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -20,11 +21,26 @@ namespace EmmerceAPIHCMUE.Controllers
         {
             try
             {
-                if (voucher.AddVoucher())
+                if (Request.Headers["Authorization"] != "")
                 {
-                    return new ResponseData(Constants.Instance.SUCCESS_CODE, Constants.Instance.ADD_SUCESS1, null);
+                    CustomMiddleware middleware = new CustomMiddleware(Request.Headers["Authorization"]);
+                    if (middleware.ValidateToken(Constants.Instance.ADMIN_ROLE1))
+                    {
+                        if (voucher.AddVoucher())
+                        {
+                            return new ResponseData(Constants.Instance.SUCCESS_CODE, Constants.Instance.ADD_SUCESS1, null);
+                        }
+                        return new ResponseData(Constants.Instance.FAIL_CODE, Constants.Instance.SOMETHING_WAS_WRONG, null);
+                    }
+                    else
+                    {
+                        return new ResponseData(Constants.Instance.FAIL_CODE, Constants.Instance.SESSION_EXPIRED, null);
+                    }
                 }
-                return new ResponseData(Constants.Instance.FAIL_CODE, Constants.Instance.SOMETHING_WAS_WRONG, null);
+                else
+                {
+                    return new ResponseData(Constants.Instance.FAIL_CODE, Constants.Instance.SESSION_EXPIRED, null);
+                }
             }
             catch (Exception e)
             {
@@ -32,16 +48,68 @@ namespace EmmerceAPIHCMUE.Controllers
             }
         }
 
+       
         [HttpPost("delete")]
         public ResponseData DeleteVouchers([FromBody] Voucher voucher)
         {
             try
             {
-                if (voucher.DeleteVoucher())
+                if (Request.Headers["Authorization"] != "")
                 {
-                    return new ResponseData(Constants.Instance.SUCCESS_CODE, Constants.Instance.DELETE_SUCESS1, null);
+                    CustomMiddleware middleware = new CustomMiddleware(Request.Headers["Authorization"]);
+                    if (middleware.ValidateToken(Constants.Instance.ADMIN_ROLE1))
+                    {
+                        if (voucher.DeleteVoucher())
+                        {
+                            return new ResponseData(Constants.Instance.SUCCESS_CODE, Constants.Instance.DELETE_SUCESS1, null);
+                        }
+                        return new ResponseData(Constants.Instance.FAIL_CODE, Constants.Instance.SOMETHING_WAS_WRONG, null);
+                    }
+                    else
+                    {
+                        return new ResponseData(Constants.Instance.FAIL_CODE, Constants.Instance.SESSION_EXPIRED, null);
+                    }
                 }
+                else
+                {
+                    return new ResponseData(Constants.Instance.FAIL_CODE, Constants.Instance.SESSION_EXPIRED, null);
+                }
+
+            }
+            catch (Exception e)
+            {
                 return new ResponseData(Constants.Instance.FAIL_CODE, Constants.Instance.SOMETHING_WAS_WRONG, null);
+            }
+        }
+
+        [HttpGet("all")]
+        public ResponseData GetAllVoucher()
+        {
+            try
+            {
+                if (Request.Headers["Authorization"] != "")
+                {
+                    CustomMiddleware middleware = new CustomMiddleware(Request.Headers["Authorization"]);
+                    if (middleware.ValidateToken(Constants.Instance.ADMIN_ROLE1))
+                    {
+                        Voucher voucher = new Voucher();
+                        DataTable dt = voucher.GetAllVoucher();
+
+                        List<Voucher> resData = new List<Voucher>();
+                        foreach (DataRow row in dt.Rows)
+                        {
+                            Voucher a = new Voucher();
+                            a.IdVoucher = row["idVoucher"].ToString();
+                            a.Price = Int32.Parse(row["price"].ToString());
+                            a.ExpiredDate = row["expiredDate"].ToString();
+                            a.IsUse = Int32.Parse(row["isUse"].ToString());
+
+                            resData.Add(a);
+                        }
+                        return new ResponseData(Constants.Instance.SUCCESS_CODE, Constants.Instance.SUCCESS_MESSAGE1, resData);
+                    }
+                }
+                return new ResponseData(Constants.Instance.FAIL_CODE, Constants.Instance.FAIL_MESSAGE1, null);
             }
             catch (Exception e)
             {
